Improved testing Mild refactoring
